modify output
Output was modified to not be divided by interval, as that analysis
could be done later if we have the original values.  Also, a .flush was
added so that everything is written to the file each interval, as
otherwise you have no data if it crashes or the run has to be killed.